requirements.txt
fastapi>=0.115
uvicorn>=0.34
opencv-python-headless>=4.10
onnxruntime>=1.17
numpy
Pillow
scipy
python-multipart
